+++ src/background/utils/updater.rs
@@ -3,7 +3,7 @@ use tauri_plugin_updater::{Update, UpdaterExt};
 
 use crate::{
     error_handler::Result,
-    modules::cli::{ServiceClient, SvcAction},
+    modules::cli::{SvcAction, TcpService},
     seelen::get_app_handle,
     state::application::FULL_STATE,
 };
@@ -55,7 +55,7 @@ pub async fn trace_update_intallation(update: Update) -> Result<()> {
             || log::trace!("Update: download finished"),
         )
         .await?;
-    ServiceClient::request(SvcAction::Stop)?;
+    TcpService::request(SvcAction::Stop)?;
     update.install(bytes)?;
     log::trace!("Update: intallation finished");
     Ok(())