+++ src/service/windows_api/app_bar.rs
@@ -1,0 +1,64 @@
+use windows::Win32::{
+    Foundation::{HWND, LPARAM},
+    UI::Shell::{
+        SHAppBarMessage, ABE_BOTTOM, ABE_LEFT, ABE_RIGHT, ABE_TOP, ABM_GETSTATE, ABM_SETSTATE,
+        ABS_ALWAYSONTOP, ABS_AUTOHIDE, APPBARDATA,
+    },
+};
+
+#[allow(dead_code)]
+pub enum AppBarDataEdge {
+    Left = ABE_LEFT as isize,
+    Top = ABE_TOP as isize,
+    Right = ABE_RIGHT as isize,
+    Bottom = ABE_BOTTOM as isize,
+}
+
+/// https://learn.microsoft.com/en-us/windows/win32/shell/abm-setstate#parameters
+#[derive(Debug, Clone, Copy)]
+pub enum AppBarDataState {
+    BothOff = 0,
+    AutoHide = ABS_AUTOHIDE as isize,
+    AlwaysOnTop = ABS_ALWAYSONTOP as isize,
+    BothOn = 3,
+}
+
+impl From<AppBarDataState> for LPARAM {
+    fn from(val: AppBarDataState) -> Self {
+        LPARAM(val as isize)
+    }
+}
+
+impl From<u32> for AppBarDataState {
+    fn from(state: u32) -> Self {
+        match state {
+            0 => AppBarDataState::BothOff,
+            ABS_AUTOHIDE => AppBarDataState::AutoHide,
+            ABS_ALWAYSONTOP => AppBarDataState::AlwaysOnTop,
+            3 => AppBarDataState::BothOn,
+            _ => unreachable!(),
+        }
+    }
+}
+
+pub struct AppBarData(pub APPBARDATA);
+impl AppBarData {
+    pub fn from_handle(hwnd: HWND) -> Self {
+        Self(APPBARDATA {
+            cbSize: std::mem::size_of::<APPBARDATA>() as u32,
+            hWnd: hwnd,
+            ..Default::default()
+        })
+    }
+
+    pub fn state(&self) -> AppBarDataState {
+        let mut data = self.0;
+        AppBarDataState::from(unsafe { SHAppBarMessage(ABM_GETSTATE, &mut data) as u32 })
+    }
+
+    pub fn set_state(&self, state: AppBarDataState) {
+        let mut data = self.0;
+        data.lParam = state.into();
+        unsafe { SHAppBarMessage(ABM_SETSTATE, &mut data) };
+    }
+}